An erroneous 'os.system' call in 'patric_2d_mad.py', overriding the users choice, was removed. The script queries the name of the computer on which it was launched.
A second marker 'SEPTUM', marking the beginning of the septum, was inserted into 'sis18_inj.mad'.
(The older one is placed at its ending.)
This way the septum is modeled more accurately.
The thickness of the septum was set to 0.1 mm.
In a few test runs, the losses were affected negligibly by the thickness but noteworthy by the length (increase > 10%).

The threshold for a warning because of a too large deflection of the bumpers in 'Bump.cpp' was raised from 6 to 8.4, following the information from D. Ondreka.

'Pic::SG' was simplified.


git-svn-id: file:///Users/sappel/Downloads/subversion/patric_sp/trunk@45 0074b501-af41-4ac2-941d-f11b2edd8d42