Chromaticity was included. 'BeamLine::read_madx_twiss' calculates the product of each quadrupoles length and deflection strength (k) from the Twiss parameters if the chromaticity flag ('chroma') is on and writes it to 'tw.xix' and 'tw.xiy', correspondingly. A division by the relativistic beta is necessary due to the MADX convention. In all other cases xix and xiy are 0. In 'SectorMap::transport' this quantity is multiplied with x or y and the momentum deviation of each particle to obtain the deflection due to chromaticity in first order. (Discussed in notes, p. 148.) The signatures of the functions involved were adapted. The original chromaticity function (applicable for constant focusing) was commented out in 'Main.cpp'. Test runs yielded reasonable tune shifts in the vertical plane but not in the horizontal one. Since the same coefficients are used (except the sign) and also without chromaticity the horizontal tune depends on the particle's momentum it is assumed that the dispersion interferes with the determination of the phase advance (which does not include dispersion).
The multiplication of MADX' 'MUX' and "MUY' with 2 pi, turning the tune to a phase advance, was moved from 'Bump::Bump' to 'BeamLine::read_madx_twiss' as it is needed to calculate the coefficients for the chromaticity kicks.

A bug in 'Bump::decrement', allowing negative deflections to occur and preventing the collapse of the bump, was removed.

The quadrupoles in 'sis18_inj.mad' were split into slices to reduce the numerical error of the chromaticity kicks.
For the full length of the quadrupoles, the linearization of the chromaticity is not justified.

Some needless functions in 'SectorMap.h' were removed.

The formerly inoperational script 'debug.sh' was fixed.
Actions meanwhile included into 'patric_2d_madx.py' were removed or disabled.
'patric_2d_madx.py' invokes make now.

'scanq.sh' was renamed 'scan.sh' as it can be employed to efficiently launch simulations to perform scans of arbitrary parameters.

A variable 'path' was introduced in 'debug.sh' for convenience.

'gdbflag=0' deactivates the waiting loop in 'Main.cpp', contrarily to the past.
'start.gdb' was adjusted to this redefinition.

The range setting in 'emittance.pro' was improved.

git-svn-id: file:///Users/sappel/Downloads/subversion/patric_sp/trunk@51 0074b501-af41-4ac2-941d-f11b2edd8d42